Use int256 / uint instead of int72 / uint64 for debt tracking
Prevents storing min_int as debt since it can not be opposed

+++ contracts/tests/TestContract.sol
@@ -13,10 +13,10 @@ contract TestContract is DebtTracking {
         int argument
     );
 
-    function increaseDebt(address creditor, uint64 value) external {
+    function increaseDebt(address creditor, uint value) external {
     }
 
-    function getDebt(address, address) public view returns (int80) {
+    function getDebt(address, address) public view returns (int) {
         return 0;
     }
